fix: repoint imports to relocated modules after rebase
Commits that replayed without conflict kept pre-reorg import paths
(config/utils/sqlsecurity, config/connectionmanager, @/testing).
