+++ frontend/tests/integration/procedures-route-authz.integration.test.ts
@@ -32,7 +32,7 @@ vi.mock('@/auth', () => ({
 
 // Route the route's ConnectionManager.executeQuery at the shared test connection
 // so authorized requests execute against the real isolated test schema.
-vi.mock('@/config/connectionmanager', () => {
+vi.mock('@/lib/db/connectionmanager', () => {
   const manager = {
     executeQuery: async (query: string, params?: unknown[]) => {
       if (!sharedState.connection) {
